first version of gui
