+++ crates/sui-cluster-test/Cargo.toml
@@ -18,6 +18,8 @@ clap = { version = "3.1.14", features = ["derive"] }
 reqwest = { version = "0.11.11", features = ["blocking", "json"] }
 async-trait = "0.1.57"
 anyhow = { version = "1.0.64", features = ["backtrace"] }
+strum_macros = "0.24.3"
+strum = "0.24.1"
 bcs = "0.1.4"
 uuid = {version = "1.1.2", features = [ "v4", "fast-rng"]}
 prometheus = "0.13.3"
