+++ crates/sui-cluster-test/src/config.rs
@@ -1,8 +1,9 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
-use clap::*;
+use anyhow::{anyhow, bail};
+use clap::{ArgEnum, Parser};
 
-#[derive(Parser, Clone, ArgEnum)]
+#[derive(Parser, Clone, ArgEnum, Debug, strum_macros::Display)]
 pub enum Env {
     Devnet,
     Staging,
@@ -12,7 +13,7 @@ pub enum Env {
     NewLocal,
 }
 
-#[derive(Parser)]
+#[derive(Parser, Clone, Debug)]
 #[clap(name = "", rename_all = "kebab-case")]
 pub struct ClusterTestOpt {
     #[clap(arg_enum)]
@@ -32,6 +33,23 @@ impl ClusterTestOpt {
             faucet_address: None,
             fullnode_address: None,
             websocket_address: None,
+        }
+    }
+}
+
+impl TryFrom<&String> for ClusterTestOpt {
+    type Error = anyhow::Error;
+    fn try_from(env: &String) -> Result<Self, Self::Error> {
+        match Env::from_str(env, true).map_err(|_| anyhow!("Failed to parse {env} as Env"))? {
+            Env::CustomRemote | Env::NewLocal => {
+                bail!("Can't parse from Env::CustomRemote | Env::NewLocal");
+            }
+            other_env => Ok(ClusterTestOpt {
+                env: other_env,
+                faucet_address: None,
+                fullnode_address: None,
+                websocket_address: None,
+            }),
         }
     }
 }
