+++ crates/sui-tool/src/bin/cli-tool.rs
@@ -6,8 +6,7 @@ use clap::*;
 use colored::Colorize;
 use sui_types::exit_main;
 
-mod commands;
-use commands::ToolCommand;
+use sui_tool::commands::ToolCommand;
 
 #[tokio::main]
 async fn main() {
