+++ crates/sui-tool/src/bin/templates/base.html.tera
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8" />
+    <title> Sui GUI Tool - {{ title }}</title>
+  </head>
+  <body>
+    {% include "nav" %}
+
+    {% block content %}{% endblock content %}
+
+  </body>
+</html>
