+++ crates/sui-tool/src/bin/templates/cluster-test.html.tera
@@ -1,0 +1,6 @@
+{% extends "base" %}
+
+{% block content %}
+    <h3>Cluster Test Response</h3>
+    <pre> {{response}} </pre>
+{% endblock content %}
