+++ crates/sui-tool/src/bin/templates/error.html.tera
@@ -1,0 +1,7 @@
+{% extends "base" %}
+
+{% block content %}
+    <h1>Error!</h1>
+    <h3>{{title}}</h3>
+    <pre>{{error}}</pre>
+{% endblock content %}
