+++ crates/sui-tool/src/bin/templates/faucet.html.tera
@@ -1,0 +1,6 @@
+{% extends "base" %}
+
+{% block content %}
+    <h3>Faucet Response</h3>
+    <pre> {{response}} </pre>
+{% endblock content %}
