+++ crates/sui-tool/src/bin/templates/nav.html.tera
@@ -1,0 +1,1 @@
+<a href="/">Home</a> | <a href="/validators">Validators</a> | <a href="/faucet">Faucet</a> | <a href="/cluster_test">Run Cluster Test</a> | <a href="/object">Object</a> | <a href="/transaction">Transaction</a>
