+++ crates/sui-tool/src/bin/templates/object.html.tera
@@ -1,0 +1,6 @@
+{% extends "base" %}
+
+{% block content %}
+    <h3>Object Response</h3>
+    <pre>{{response}})</pre>
+{% endblock content %}
