+++ crates/sui-tool/src/bin/templates/transaction.html.tera
@@ -1,0 +1,6 @@
+{% extends "base" %}
+
+{% block content %}
+    <h3>Transaction Response</h3>
+    <pre>{{response}})</pre>
+{% endblock content %}
